Add dashboard rendering mock data
**Problem**
It is difficult to know how to structure the metrics generated from the
node modules without having a visual overview of how the data will be
displayed.

**Solution**
Create a front-end dashboard which renders the metrics into a table
generated using the Datatable library.

The metrics object is stored as a javascript in the `repos-metrics.js`
file. It is then being parsed into a flat array to be sent to the table.

Some meta data are used to dynamically set some DOM element, like the
period duration and the organization name.

**Result**
The beta version of the dashboard can be seen loading the `index.html`
file into a browser. There is not require dependencies.