Cache and get pull requests details for Github org
**Problem**
Not possible get all pull requests detailed information for all
repositories of an organization and cache it on the file system.

**Solution**
Implement a function `createCache` that query all organization
repositories to get pull requests details (including commits and
comments) and cache it on the file system.

The cache system is one file per pull requests. When doing a query to
`createCache` all pull requests are checked to see if they exist on
dist.

**Result**
Use the function `createCache` which return a promise with all pull
requests details.